Safari-611.1.10/LayoutTests/platform/gtk/fast/canvas/toDataURL-supportedTypes-expected.txt

 The Actual Canvas
Given MIMEType: image/png
Used MIMEType: image/png
MIME types are the SAME.

Given MIMEType: image/jpeg
Used MIMEType: image/jpeg
MIME types are the SAME.

Given MIMEType: image/gif
Used MIMEType: image/png
MIME types DIFFER.

Given MIMEType:
Used MIMEType: image/png
MIME types DIFFER.

Given MIMEType: null
Used MIMEType: image/png
MIME types DIFFER.

Given MIMEType: undefined
Used MIMEType: image/png
MIME types DIFFER.

Given MIMEType: image/x-webkitbitmap
Used MIMEType: image/png
MIME types DIFFER.
